+++ modules/network-firewall-policy/versions.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2023 Google LLC
+ * Copyright 2025 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,20 +15,20 @@
  */
 
 terraform {
-  required_version = ">= 1.3.0"
+  required_version = ">= 1.3"
 
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 4.64, < 6"
+      version = ">= 6.28, < 8"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 4.64, < 6"
+      version = ">= 6.28, < 8"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network:network-firewall-policy/v9.1.0"
+    module_name = "blueprints/terraform/terraform-google-network:network-firewall-policy/v12.0.0"
   }
 }