+++ modules/network-peering/versions.tf
@@ -15,16 +15,16 @@
  */
 
 terraform {
-  required_version = ">= 0.13.0"
+  required_version = ">= 1.3"
 
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 4.65, < 6"
+      version = ">= 4.65, < 8"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 4.65, < 6"
+      version = ">= 4.65, < 8"
     }
     random = {
       source  = "hashicorp/random"
@@ -37,6 +37,6 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network:network-peering/v9.1.0"
+    module_name = "blueprints/terraform/terraform-google-network:network-peering/v12.0.0"
   }
 }