+++ modules/private-service-connect/versions.tf
@@ -15,23 +15,23 @@
  */
 
 terraform {
-  required_version = ">= 0.13"
+  required_version = ">= 1.3"
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 5.8, < 6"
+      version = ">= 5.8, < 8"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 5.8, < 6"
+      version = ">= 5.8, < 8"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network:private-service-connect/v9.1.0"
+    module_name = "blueprints/terraform/terraform-google-network:private-service-connect/v12.0.0"
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-network:private-service-connect/v9.1.0"
+    module_name = "blueprints/terraform/terraform-google-network:private-service-connect/v12.0.0"
   }
 }