+++ modules/routes-beta/versions.tf
@@ -15,16 +15,16 @@
  */
 
 terraform {
-  required_version = ">= 0.13.0"
+  required_version = ">= 1.3"
 
   required_providers {
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 3.0, < 6"
+      version = ">= 3.0, < 8"
     }
   }
 
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-network:routes-beta/v9.1.0"
+    module_name = "blueprints/terraform/terraform-google-network:routes-beta/v12.0.0"
   }
 }