+++ modules/routes/versions.tf
@@ -15,16 +15,16 @@
  */
 
 terraform {
-  required_version = ">= 0.13.0"
+  required_version = ">= 1.3"
 
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = ">= 3.83, < 6"
+      version = ">= 3.83, < 8"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-network:routes/v9.1.0"
+    module_name = "blueprints/terraform/terraform-google-network:routes/v12.0.0"
   }
 }